Struct proptest_ext::ValueGenerator

pub struct ValueGenerator { /* private fields */ }

Context for generating single values out of strategies.

Proptest is designed to be built around “value trees”, which represent a spectrum from complex values to simpler ones. But in some contexts, like benchmarking or generating corpuses, one just wants a single value. This is a convenience struct for that.

Implementations

impl ValueGenerator

pub fn new() -> Self

Creates a new value generator with the default RNG.

pub fn deterministic() -> Self

Creates a new value generator with a deterministic RNG.

This generator has a hardcoded seed, so its results are predictable across test runs. However, a new proptest version may change the seed.

pub fn from_seed(seed: impl Hash) -> Self

Creates a new value generator from the given seed.

This generator is typically used with a hardcoded seed that is keyed on the input data somehow. For example, for a test fixture it may be the name of the fixture.

pub fn partial_clone(&mut self) -> Self

Does a “partial clone” of the ValueGenerator, creating a new independent but deterministic RNG.

pub fn generate<S: Strategy>(&mut self, strategy: S) -> S::Value

Generates a single value for this strategy.

Panics if generating the new value fails. The only situation in which this can happen is if generating the value causes too many internal rejects.